Convert ColorCycle to render callback with context
